refactor: add type for readability
